[Refactor] terrain-actions-table.cpp/h をterrain-definition.cpp/h に吸収合併した
