+++ VisualStudio/Hengband/Hengband.vcxproj.filters
@@ -2544,9 +2544,6 @@
     <ClCompile Include="..\..\src\system\floor\town-list.cpp">
       <Filter>system\floor</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\system\terrain\terrain-actions-table.cpp">
-      <Filter>system\terrain</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\system\terrain\terrain-definition.cpp">
       <Filter>system\terrain</Filter>
     </ClCompile>
@@ -5516,9 +5513,6 @@
     </ClInclude>
     <ClInclude Include="..\..\src\game-option\game-option-page.h">
       <Filter>game-option</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\system\terrain\terrain-actions-table.h">
-      <Filter>system\terrain</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\system\terrain\terrain-definition.h">
       <Filter>system\terrain</Filter>
